Development to master for 2.3.9.2

+++ all_in_one_seo_pack.php
@@ -3,7 +3,7 @@
 Plugin Name: All In One SEO Pack
 Plugin URI: http://semperfiwebdesign.com
 Description: Out-of-the-box SEO for your WordPress blog. Features like XML Sitemaps, SEO for custom post types, SEO for blogs or business sites, SEO for ecommerce sites, and much more. Almost 30 million downloads since 2007.
-Version: 2.3.9
+Version: 2.3.9.2
 Author: Michael Torbert
 Author URI: http://michaeltorbert.com
 Text Domain: all-in-one-seo-pack
@@ -31,14 +31,14 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
  * The original WordPress SEO plugin.
  *
  * @package All-in-One-SEO-Pack
- * @version 2.3.9
+ * @version 2.3.9.2
  */
 
 if ( ! defined( 'AIOSEOPPRO' ) ) {
 	define( 'AIOSEOPPRO', false );
 }
 if ( ! defined( 'AIOSEOP_VERSION' ) ) {
-	define( 'AIOSEOP_VERSION', '2.3.9' );
+	define( 'AIOSEOP_VERSION', '2.3.9.2' );
 }
 global $aioseop_plugin_name;
 $aioseop_plugin_name = 'All in One SEO Pack';
@@ -247,6 +247,10 @@ if ( ! function_exists( 'aioseop_activate' ) ) {
 		}
 		$aiosp_activation = true;
 
+		if( ! is_network_admin() || !isset( $_GET['activate-multi'] ) ) {
+			set_transient( '_aioseop_activation_redirect', true, 30 ); // Sets 30 second transient for welcome screen redirect on activation.
+			}
+
 		delete_user_meta( get_current_user_id(), 'aioseop_yst_detected_notice_dismissed' );
 
 		if ( AIOSEOPPRO ) {
@@ -384,6 +388,10 @@ if ( ! function_exists( 'aioseop_init_class' ) ) {
 		require_once( AIOSEOP_PLUGIN_DIR . 'public/opengraph.php' );
 		require_once( AIOSEOP_PLUGIN_DIR . 'inc/compatability/compat-init.php');
 		require_once( AIOSEOP_PLUGIN_DIR . 'public/front.php' );
+		require_once( AIOSEOP_PLUGIN_DIR . 'public/google-analytics.php' );
+		require_once( AIOSEOP_PLUGIN_DIR . 'admin/display/welcome.php' );
+
+		$aioseop_welcome = new aioseop_welcome(); // TODO move this to updates file.
 
 		if ( AIOSEOPPRO ) {
 			require_once( AIOSEOP_PLUGIN_DIR . 'pro/class-aio-pro-init.php' ); // Loads pro files and other pro init stuff.
@@ -398,6 +406,8 @@ if ( ! function_exists( 'aioseop_init_class' ) ) {
 			$aioseop_pro_updates = new AIOSEOP_Pro_Updates();
 			add_action( 'admin_init', array( $aioseop_pro_updates, 'version_updates' ), 12 );
 		}
+
+		add_action( 'admin_init', 'aioseop_welcome' );
 
 		if ( aioseop_option_isset( 'aiosp_unprotect_meta' ) ) {
 			add_filter( 'is_protected_meta', 'aioseop_unprotect_meta', 10, 3 );
@@ -416,6 +426,19 @@ if ( ! function_exists( 'aioseop_init_class' ) ) {
 				$current_screen = WP_Screen::get( 'front' );
 			}
 		}
+	}
+}
+
+
+
+if ( ! function_exists( 'aioseop_welcome' ) ){
+	function aioseop_welcome(){
+		if( get_transient( '_aioseop_activation_redirect') ){
+			$aioseop_welcome = new aioseop_welcome();
+			delete_transient( '_aioseop_activation_redirect' );
+			$aioseop_welcome->init();
+		}
+
 	}
 }
